apply digits2 layer
