+++ src/areacode/features/map/index.tsx
@@ -18,21 +18,32 @@ import { MACompListContent } from 'areacode/pages/list/MACompListContent'
 import { MAAreaCodeInfoCards } from 'areacode/pages/list/components'
 import { getColorStyleByAreaCode } from 'areacode/components'
 
-const MA_MAP_STYLE =
-  'https://tile.openstreetmap.jp/styles/osm-bright-ja/style.json'
+// const MA_MAP_STYLE =
+//   'https://tile.openstreetmap.jp/styles/osm-bright-ja/style.json'
 
 type PopupInfoType = {
   longitude: number
   latitude: number
   properties: Record<string, string>
 }
+
+type HoverState = {
+  sourceId: 'ma-source' | 'digits2-source'
+  featureId: string | number
+}
+
 function App() {
-  const [geoData, setGeoData] = useState<FeatureCollection<Geometry> | null>(
-    null,
-  )
+  const [maGeoData, setMaGeoData] =
+    useState<FeatureCollection<Geometry> | null>(null)
+  const [digits2GeoData, setDigits2GeoData] =
+    useState<FeatureCollection<Geometry> | null>(null)
+  const [showMA, setShowMA] = useState(true)
+  const [showDigits2, setShowDigits2] = useState(true)
+
   const [popupInfo, setPopupInfo] = useState<PopupInfoType | null>(null)
   const mapRef = useRef<MapLibreMap | null>(null)
-  const hoverIdRef = useRef<string | number | null>(null)
+
+  const hoverRef = useRef<HoverState | null>(null)
 
   useEffect(() => {
     fetch('/map/areacode.json')
@@ -54,54 +65,84 @@ function App() {
             },
           }
         })
-        setGeoData(geojson)
+        setMaGeoData(geojson)
+      })
+  }, [])
+
+  useEffect(() => {
+    fetch('/map/digits2.json')
+      // GeoJSON is simplified to TopoJSON on https://mapshaper.org/
+      .then((res) => res.json())
+      .then((topoData) => {
+        const geojson = topojson.feature(
+          topoData,
+          topoData.objects['2digits'],
+        ) as unknown as FeatureCollection<Geometry>
+        geojson.features = geojson.features.map((f: Feature<Geometry>) => {
+          const properties = (f.properties ?? {}) as Record<string, string>
+          return {
+            ...f,
+            properties: {
+              ...properties,
+              fillColor: getColorStyleByAreaCode(`0${properties['_市外局番']}`)
+                .background.backgroundColor,
+            },
+          }
+        })
+        setDigits2GeoData(geojson)
       })
   }, [])
 
   const clearHoverState = useCallback(() => {
-    if (!mapRef.current || hoverIdRef.current === null) {
+    if (!mapRef.current || hoverRef.current === null) {
       return
     }
 
     mapRef.current.setFeatureState(
-      { source: 'ma-source', id: hoverIdRef.current },
+      { source: hoverRef.current.sourceId, id: hoverRef.current.featureId },
       { hover: false },
     )
-    hoverIdRef.current = null
+    hoverRef.current = null
   }, [])
 
-  const onClick = (event: MapLayerMouseEvent) => {
-    const feature = event.features && event.features[0]
-    if (feature) {
-      setPopupInfo({
-        longitude: event.lngLat.lng,
-        latitude: event.lngLat.lat,
-        properties: feature.properties,
-      })
+  const onClick = useCallback((event: MapLayerMouseEvent) => {
+    const feature = event.features?.[0]
+    if (!feature || feature.source !== 'ma-source' || !feature.properties) {
+      setPopupInfo(null)
+      return
     }
-  }
+
+    setPopupInfo({
+      longitude: event.lngLat.lng,
+      latitude: event.lngLat.lat,
+      properties: feature.properties as Record<string, string>,
+    })
+  }, [])
 
   const onHover = useCallback(
     (event: MapLayerMouseEvent) => {
       mapRef.current = event.target
       const feature = event.features?.[0]
       const nextHoverId = feature?.id
+      const sourceId = feature?.source as HoverState['sourceId'] | undefined
 
-      if (nextHoverId === undefined) {
+      if (nextHoverId === undefined || !sourceId) {
         clearHoverState()
         return
       }
-
-      if (hoverIdRef.current === nextHoverId) {
+      if (
+        hoverRef.current?.sourceId === sourceId &&
+        hoverRef.current.featureId === nextHoverId
+      ) {
         return
       }
 
       clearHoverState()
       event.target.setFeatureState(
-        { source: 'ma-source', id: nextHoverId },
+        { source: sourceId, id: nextHoverId },
         { hover: true },
       )
-      hoverIdRef.current = nextHoverId
+      hoverRef.current = { sourceId, featureId: nextHoverId }
     },
     [clearHoverState],
   )
@@ -112,37 +153,7 @@ function App() {
     }
   }, [clearHoverState])
 
-  const borderStyle: LineLayerSpecification = {
-    source: 'ma-source',
-    id: 'ma-borders',
-    type: 'line',
-    paint: {
-      'line-color': '#1a1a1a',
-      'line-width': ['interpolate', ['linear'], ['zoom'], 4, 0.4, 8, 1.2],
-      'line-opacity': 0.8,
-    },
-  }
-
-  const labelStyle: SymbolLayerSpecification = {
-    source: 'ma-source',
-    id: 'ma-labels',
-    type: 'symbol',
-    layout: {
-      'text-field': ['coalesce', ['get', '_MA名']],
-      'text-size': ['interpolate', ['linear'], ['zoom'], 5, 9, 8, 12],
-      'text-font': ['Noto Sans Regular'],
-      'text-allow-overlap': false,
-      'text-ignore-placement': false,
-    },
-    paint: {
-      'text-color': '#111827',
-      'text-halo-color': '#ffffff',
-      'text-halo-width': 1.2,
-      'text-halo-blur': 0.2,
-    },
-  }
-
-  const fillStyle: FillLayerSpecification = {
+  const maFillStyle: FillLayerSpecification = {
     source: 'ma-source',
     id: 'ma-fills',
     type: 'fill',
@@ -152,37 +163,148 @@ function App() {
       'fill-opacity': [
         'case',
         ['boolean', ['feature-state', 'hover'], false],
-        0.5,
-        1.0,
+        0.55,
+        0.85,
       ],
     },
   }
 
+  const maBorderStyle: LineLayerSpecification = {
+    source: 'ma-source',
+    id: 'ma-borders',
+    type: 'line',
+    paint: {
+      'line-color': '#1a1a1a',
+      'line-width': ['interpolate', ['linear'], ['zoom'], 4, 0.4, 8, 1.2],
+      'line-opacity': 0.7,
+    },
+  }
+
+  const maLabelStyle: SymbolLayerSpecification = {
+    source: 'ma-source',
+    id: 'ma-labels',
+    type: 'symbol',
+    layout: {
+      'text-field': ['coalesce', ['get', '_MA名']],
+      'text-size': ['interpolate', ['linear'], ['zoom'], 5, 8, 8, 11],
+      'text-font': ['Noto Sans Regular'],
+      'text-allow-overlap': false,
+      'text-ignore-placement': false,
+    },
+    paint: {
+      'text-color': '#111827',
+      'text-halo-color': '#ffffff',
+      'text-halo-width': 1.1,
+      'text-halo-blur': 0.2,
+    },
+  }
+
+  const digits2BorderStyle: LineLayerSpecification = {
+    source: 'digits2-source',
+    id: 'digits2-borders',
+    type: 'line',
+    paint: {
+      'line-color': '#374151',
+      'line-width': ['interpolate', ['linear'], ['zoom'], 4, 0.6, 8, 2.0],
+      'line-opacity': 0.85,
+    },
+  }
+
+  const digits2LabelStyle: SymbolLayerSpecification = {
+    source: 'digits2-source',
+    id: 'digits2-labels',
+    type: 'symbol',
+    layout: {
+      'text-field': ['coalesce', ['get', '市外局番2桁']],
+      'text-size': ['interpolate', ['linear'], ['zoom'], 5, 20, 8, 30],
+      'text-font': ['Noto Sans Bold'],
+      'text-allow-overlap': false,
+      'text-ignore-placement': false,
+    },
+    paint: {
+      'text-color': '#111827',
+      'text-halo-color': '#ffffff',
+      'text-halo-width': 1.3,
+      'text-halo-blur': 0.2,
+    },
+  }
+
+  const interactiveLayerIds = [
+    ...(showMA ? ['ma-fills'] : []),
+    ...(showDigits2 ? ['digits2-fills'] : []),
+  ]
+
   return (
-    <Map
-      initialViewState={{
-        longitude: 139.76711,
-        latitude: 35.68074,
-        zoom: 6,
-      }}
-      style={{ position: 'absolute', top: 0, bottom: 0, width: '100%' }}
-      // mapStyle={MA_MAP_STYLE}
-      onClick={onClick}
-      onMouseMove={onHover}
-      onMouseLeave={clearHoverState}
-      interactiveLayerIds={['ma-fills']}
-    >
-      {geoData && (
-        <Source id="ma-source" type="geojson" data={geoData} generateId={true}>
-          <Layer {...fillStyle}></Layer>
-          <Layer {...borderStyle}></Layer>
-          <Layer {...labelStyle}></Layer>
-        </Source>
-      )}
-      {popupInfo && (
-        <PopupInfo info={popupInfo} setFunc={setPopupInfo}></PopupInfo>
-      )}
-    </Map>
+    <div style={{ position: 'relative', width: '100vw', height: '100vh' }}>
+      <div
+        style={{
+          position: 'absolute',
+          zIndex: 1,
+          top: 12,
+          left: 12,
+          padding: '8px 12px',
+          backgroundColor: 'rgba(255,255,255,0.9)',
+          borderRadius: 8,
+          boxShadow: '0 2px 8px rgba(0,0,0,0.15)',
+          fontSize: 14,
+        }}
+      >
+        <label style={{ display: 'block', marginBottom: 4 }}>
+          <input
+            type="checkbox"
+            checked={showMA}
+            onChange={(event) => setShowMA(event.target.checked)}
+          />{' '}
+          MA地図
+        </label>
+        <label style={{ display: 'block' }}>
+          <input
+            type="checkbox"
+            checked={showDigits2}
+            onChange={(event) => setShowDigits2(event.target.checked)}
+          />{' '}
+          市外局番2桁地図
+        </label>
+      </div>
+      <Map
+        initialViewState={{
+          longitude: 139.76711,
+          latitude: 35.68074,
+          zoom: 6,
+        }}
+        style={{ position: 'absolute', top: 0, bottom: 0, width: '100%' }}
+        // mapStyle={MA_MAP_STYLE}
+        onClick={onClick}
+        onMouseMove={onHover}
+        onMouseLeave={clearHoverState}
+        interactiveLayerIds={interactiveLayerIds}
+      >
+        {showDigits2 && digits2GeoData && (
+          <Source
+            id="digits2-source"
+            type="geojson"
+            data={digits2GeoData}
+            generateId={true}
+          >
+            <Layer {...digits2BorderStyle}></Layer>
+            <Layer {...digits2LabelStyle}></Layer>
+          </Source>
+        )}
+        {showMA && maGeoData && (
+          <Source
+            id="ma-source"
+            type="geojson"
+            data={maGeoData}
+            generateId={true}
+          >
+            <Layer {...maFillStyle}></Layer>
+            <Layer {...maBorderStyle}></Layer>
+            <Layer {...maLabelStyle}></Layer>
+          </Source>
+        )}
+        {popupInfo && <PopupInfo info={popupInfo} setFunc={setPopupInfo} />}
+      </Map>
+    </div>
   )
 }
